Refactor to have api calls in one file
