+++ src/auth/Signup.js
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
-import registerUserAction from '../actions/registerUser';
+import { registerUserAction } from '../api/api';
 // import Login from './Login';
 import './Login.css';
